requirements-heroku.txt
# Requirements para Heroku - Arcane AI SaaS

# Base requirements
-r requirements.txt

# Heroku specific
gunicorn==21.2.0
gevent==23.9.1
whitenoise==6.6.0
dj-database-url==2.1.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# Cache
django-redis==5.4.0
redis==5.0.1

# Security
django-cors-headers==4.3.1
django-ratelimit==4.1.0

# Monitoring
sentry-sdk==1.38.0

# Email
sendgrid==6.10.0

# Payment processing
stripe==7.8.0

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0

# API
djangorestframework==3.14.0
django-filter==23.5

# File processing
Pillow==10.1.0

# Utilities
python-slugify==8.0.1

# AI/ML (lightweight versions for Heroku)
openai==1.3.7
numpy==1.24.3
scikit-learn==1.3.2
faiss-cpu==1.7.4
